refactor(web): extract useCharacters hook from App.jsx (sc-1651)
First slice of the App.jsx god-module decomposition (~2300 lines, ~35 useState,
~60 functions). Moves the character roster state + all 17 character
CRUD/reference/look/LoRA functions (withCharacterApi + create/update/archive,
references, looks, loras, test job) and refreshCharacters into
hooks/useCharacters.js. App composes the hook and passes the same values down, so
no screen prop signatures change — CharacterStudio/ImageStudio/VideoStudio are
untouched. Shared concerns (token, activeProject, setError, requestedGpu,
setActiveView, refreshData) are passed in. The sendCharacterToImage/Video
navigation helpers stay in App (they're studio-launch, not character data).

Behavior-preserving: App.jsx 2334 -> 2186 lines. 107 web vitest pass, production
build clean, and a browser smoke (Rust API on a temp data dir) confirmed creating
a character end-to-end through the hook with no console errors.
